refactored dropup menu with dynamic loading of APP and non-APP loading of 2 new components
